Update modelling.qmd
Reworked to include phase information for a cell (which parent a marker came from), added some plots. Next will be to fix recombination which as coded can move part of the same segment multiple times (each recombination event in one cell division chooses start and end of the segment afresh).